Move history query contracts under cqrs
